CarAPI refactored
